moved duplicator to util
